LU-2022 utils: Mount MGS lacking index
The MGT is the only device that can't be formatted with a
index but mount.lustre expects all devices to be formated
with a index. This patch allows the MGT to be the exception.

+++ lustre/utils/mount_lustre.c
@@ -311,7 +311,7 @@ static int parse_ldd(char *source, struct mount_opts *mop, char *options)
 		return rc;
 	}
 
-	if (ldd->ldd_flags & LDD_F_NEED_INDEX) {
+	if (!IS_MGS(ldd) && ldd->ldd_flags & LDD_F_NEED_INDEX) {
 		fprintf(stderr, "%s: %s has no index assigned "
 			"(probably formatted with old mkfs)\n",
 			progname, source);
